ci: fix psalm null reference error

+++ src/Bundle/PrimeIndexerTestBundle.php
@@ -9,15 +9,23 @@ final class PrimeIndexerTestBundle extends Bundle
 {
     public function boot(): void
     {
+        if (!$container = $this->container) {
+            return;
+        }
+
         /** @var TestingIndexer $testingIndexer */
-        $testingIndexer = $this->container->get(TestingIndexer::class);
+        $testingIndexer = $container->get(TestingIndexer::class);
         $testingIndexer->init();
     }
 
     public function shutdown(): void
     {
+        if (!$container = $this->container) {
+            return;
+        }
+
         /** @var TestingIndexer $testingIndexer */
-        $testingIndexer = $this->container->get(TestingIndexer::class);
+        $testingIndexer = $container->get(TestingIndexer::class);
         $testingIndexer->destroy();
     }
 }
